backend/python/plugins/nps/pyproject.toml
[tool.poetry]
name = "nps"
version = "0.1.0"
description = ""
authors = ["Florent Chenebault <florent.chenebault@lifen.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
pydevlake = { path = "../../pydevlake", develop = true }
google-api-python-client = "^2.133.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
